Add class for documents
This extends the parser for available documents. It is now able to guess
a filename from the documents properties which allows the downloader to
better detect already existing documents (not in use yet). Unfortunately
adpworld.de does not follow the HTTP standard and only allows GET
requests and no HEAD requests to get the real filename.