Add CLI arg to allow download of multiple files with the same name
As reported on https://github.com/nicksinger/adp_downloader/issues/11,
there is the possibility that multiple documents published on the same
date have the same name even when the contents are different.

This commit adds the `--download-all` CLI argument to force
`downloader.py` to skip the check of downloaded file names and instead
download even those documents with file names which are the same of
documents already downloaded. To avoid overwriting already downloaded
documents, it will check for the existence of a file matching the name,
and in that case add a numeric index right before the file extension.

Without `--download-all` it will work as before.

As a consequence of using `argparse`, a small help was automaticall
added:

```
$ python3 downloader.py -h
usage: downloader.py [-h] [--download-all]

adpworld.de scraper

optional arguments:
  -h, --help      show this help message and exit
  --download-all  Download all files, including those with filenames matching
                  already downloaded files
```
Normal use:
```
$ python3 downloader.py >/dev/null
$ ls downloads/|wc -l
167
```
Use with `--download-all`:
```
$ python3 downloader.py --download-all >/dev/null
$ ls downloads/|wc -l
170
```